+++ messages/ja/views_translate_index.php
@@ -1,13 +1,12 @@
 <?php
-
-return [
-    'Search' => '検索',
-    'Translation Editor' => '翻訳エディタ',
-    '<strong>Translation</strong> Editor' => '',
-    'For more informations about translation syntax see' => '',
-    'If the value is empty, the message is considered as not translated.' => '',
-    'Message string can be used with plural forms format. Check i18n section of the documentation for details.' => '',
-    'Messages that no longer need translation will have their translations enclosed between a pair of "@@" marks.' => '',
-    'Original (en-US)' => '',
-    'Translated' => '',
-];
+return array (
+  '<strong>Translation</strong> Editor' => '<strong>翻訳</strong> 編集',
+  'For more informations about translation syntax see' => '変換構文の詳細については、次を参照してください。',
+  'If the value is empty, the message is considered as not translated.' => '値が空の場合、メッセージは翻訳されていないと見なされます。',
+  'Message string can be used with plural forms format. Check i18n section of the documentation for details.' => 'メッセージ文字列は複数形で使用できます。詳細については、ドキュメントの i18n セクションを確認してください。',
+  'Messages that no longer need translation will have their translations enclosed between a pair of "@@" marks.' => '翻訳が不要になったメッセージは、翻訳が "@@" マークのペアで囲まれます。',
+  'Original (en-US)' => 'オリジナル (en-US)',
+  'Search' => '検索',
+  'Translated' => '翻訳済み',
+  'Translation Editor' => '翻訳エディタ',
+);