+++ models/Archive.php
@@ -23,22 +23,22 @@ class Archive
 
     public static function validateBasePath($lang)
     {
-        return is_dir(static::getBasePath($lang));
+        return is_dir((string)static::getBasePath($lang));
     }
 
     public static function getBasePath($lang)
     {
-        return Yii::getAlias(static::$basePath . DIRECTORY_SEPARATOR .$lang);
+        return Yii::getAlias(static::$basePath . DIRECTORY_SEPARATOR . $lang);
     }
 
     public static function getFilePath($lang)
     {
-        return Yii::getAlias(static::$basePath . DIRECTORY_SEPARATOR. $lang . DIRECTORY_SEPARATOR . 'archive.json');
+        return Yii::getAlias(static::$basePath . DIRECTORY_SEPARATOR . $lang . DIRECTORY_SEPARATOR . 'archive.json');
     }
 
     public static function update($lang, $messages)
     {
-        if(!self::validateBasePath($lang)) {
+        if (!self::validateBasePath($lang)) {
             return false;
         }