+++ docs/CHANGELOG.md
@@ -9,7 +9,11 @@ Changelog
 - Enh #40: Support attribute “target” for link tag
 - Fix #44: Allow snake case for URI vars in href
 - Fix #47: Fix tests
-
+- Enh #58: Tests for `next` version
+- Fix #59: Fix visibility of the method `Controller::getAccessRules()`
+- Fix #63: Don't purify the char `&`
+- Fix #65: Fix keeping of target in links of translated text
+- Fix #41: Allow to translate module from folder name different than module id
 
 1.0.1 (Unreleased)
 ------------------
@@ -31,4 +35,3 @@ Changelog
 - Chg: Updated HumHub min version to 1.5
 - Fix: Invalid validation error when parsing plural or selection message pattern
 - Enh #29: Added "Only show missing translation" filter toggle to translation editor
-